Border-radius corrections
Changed the border radius of the 'style2.css' file.  Looks like the anchor tags are working correctly if information is input in the text box, but the JavaScript is probably going to need correcting on my behalf.  Hadn't dealt with form/error-handling just yet, but this is something that we can move on to the next time we meet up.  But it does look like everything else is working fine.